custom_xtts/requirements.txt
transformers[torch]
datasets
accelerate
deepspeed
huggingface_hub
torchaudio
librosa
soundfile
evaluate